+++ Controller/VerificationController.php
@@ -1,0 +1,14 @@
+<?php
+namespace Publero\AppleMobileBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+
+class VerificationController extends Controller
+{
+    public function indexAction()
+    {
+        
+        return $this->render('VendorNameBundleNameBundle    :ControllerName:index.html.twig');
+    }
+
+}