+++ changelog.md
@@ -1,4 +1,10 @@
-### 0.2.0  (2012-09-xx)
+### 0.2.1
+
+* Composer and travis-ci fix
+* Store Receipt is optional
+* Additional documentation
+
+### 0.2.0
 
 * TransactionId is unique and works as @ORM\Id
 * Added more unit tests
@@ -8,6 +14,6 @@
 * Updated documentation
 * Added composer.json
 
-### 0.1.0  (2012-04-25)
+### 0.1.0 (2012-04-25)
 
 * Initial release